Add Theme\Setup to bootstrap Timber
